Start tests with multiple docker containers running at the same time
